+++ src/lib.rs
@@ -1,6 +1,7 @@
 pub(crate) mod api;
 pub(crate) mod background;
 pub mod cli;
+pub mod commercial;
 pub mod config;
 pub mod db;
 pub(crate) mod domain;